Separate views from components. Follow the same project structure as in kerrokantasi
