+++ src/scripts/index.js
@@ -2,10 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Router, Route, IndexRoute } from 'react-router';
 
-// Pages
-import App from './app.js';
-import EditorPage from './components/editorpage';
-import SearchPage from './components/searchpage/searchpage.js';
+// Views
+import App from './views/app';
+import Editor from './views/editor';
+import Search from './views/search';
 
 // Initialize tap event plugin (used by material-ui components)
 import injectTapEventPlugin from 'react-tap-event-plugin';
@@ -15,8 +15,8 @@ injectTapEventPlugin();
 ReactDOM.render(
     <Router>
         <Route path="/" component={App}>
-            <IndexRoute component={SearchPage}/>
-            <Route path="event/:action/:eventId" component={EditorPage}/>
+            <IndexRoute component={Search}/>
+            <Route path="event/:action/:eventId" component={Editor}/>
         </Route>
     </Router>,
     document.getElementById('content')
