+++ src/scripts/views/editor/index.js
